adapt benchmarks, move ProcessAPIData to api v1
